# openapi.yaml
openapi: 3.0.3
info:
  title: Taxi Service Management API
  description: |
    API REST completa para sistema de gestión de taxis multi-tenant.

    ## Características principales:
    - 🚕 Gestión de conductores y vehículos
    - 📱 Sistema de solicitud y asignación de viajes
    - 👥 Gestión de clientes
    - 💰 Sistema de tarifas dinámicas
    - ⭐ Calificaciones y reseñas
    - 🎟️ Códigos promocionales
    - 📅 Horarios de conductores
    - 🗺️ Tracking en tiempo real

    ## Autenticación:
    Todas las rutas protegidas requieren un token JWT en el header Authorization:
    ```
    Authorization: Bearer <token>
    ```
  version: 1.0.0
  contact:
    name: Taxi System API Support
    email: support@taxisystem.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.taxisystem.com/api/v1
    description: Servidor de producción
  - url: https://staging-api.taxisystem.com/api/v1
    description: Servidor de staging
  - url: http://localhost:8000/api/v1
    description: Servidor de desarrollo local
...
security:
  - BearerAuth: []
paths:
...
  /customers:
    get:
      tags: [Customers]
      summary: Listar clientes
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: customer_type
          in: query
          schema:
            type: string
            enum: [new, frequent, vip, corporate]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  customers:
                    type: array
                    items:
                      $ref: '#/components/schemas/Customer'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags: [Customers]
      summary: Crear cliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerInput'
      responses:
        '201':
          description: Cliente creado
  /customers/{customer_id}:
    get:
      tags: [Customers]
      summary: Obtener cliente por ID
      parameters:
        - $ref: '#/components/parameters/CustomerId'
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: Datos del cliente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'

    put:
      tags: [Customers]
      summary: Actualizar cliente
      parameters:
        - $ref: '#/components/parameters/CustomerId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerInput'
      responses:
        '200':
          description: Cliente actualizado
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantId:
      name: tenant_id
      in: query
      required: true
      schema:
        type: string
        example: "tenant_taxi_001"
      description: ID del tenant multi-tenant
...
    CustomerId:
      name: customer_id
      in: path
      required: true
      schema:
        type: string
      description: ID del cliente
  schemas:
...
    Customer:
      type: object
      properties:
        customer_id:
          type: string
        tenant_id:
          type: string
        phone:
          type: string
        name:
          type: string
        email:
          type: string
        profile_photo_url:
          type: string
        preferred_payment_method:
          type: string
          enum: [cash, card, transfer, wallet, corporate]
        total_rides:
          type: integer
        total_spent:
          type: number
        customer_type:
          type: string
          enum: [new, frequent, vip, corporate]
        loyalty_points:
          type: integer
        active:
          type: boolean
    CustomerInput:
      type: object
      required: [tenant_id, phone, name]
      properties:
        tenant_id:
          type: string
        phone:
          type: string
        name:
          type: string
        email:
          type: string
        password:
          type: string
        preferred_payment_method:
          type: string

    # =============================================================================
    # VEHICLE SCHEMAS
    # =============================================================================
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
...